fix: harden lifecycle command edge cases
Preserve explicit session selection and failed composer sends, retain live entries during authoritative reloads, widen compaction metrics, and clarify lifecycle timeout recovery.


Copilot-Session: c41e789a-73e3-41ae-ac62-1ff5287599c0

+++ src/OpenClaw.Shared/Sessions/SessionCreateModels.cs
@@ -38,7 +38,7 @@ public sealed class SessionCompactResult
     public string? Key { get; init; }
     public bool Compacted { get; init; }
     public string? Reason { get; init; }
-    public int? TokensBefore { get; init; }
-    public int? TokensAfter { get; init; }
+    public long? TokensBefore { get; init; }
+    public long? TokensAfter { get; init; }
     public string? Error { get; init; }
 }
